Init files
Took the BrownEdu ISP/GISP form and parameterized the parts that call for a student response, leaving the commentary in raw form.

Uses a /Sec/Week/Template for each of the 14 Weeks.  The Template could be expanded to include parameters for reports from each week.

Parameter names are very rough.